Update util.js
I was struggling with an issue on the socket.io-client on a mobile using the native android browser or internal webviews.
The stock browser, doesn't implement websockets, so it fallbacks to XHR.




The 100ms are fine for desktops, but on a mobile you still need to face with an infinite spinning wheel. I've incremented the defer timeout to 500ms and the problem was gonne! 




P.S:
This issue doesn't occur using chrome in android because it already implements websockets.